+++ src/RestApi/Device/Telemetry/DeviceTelemetryManagement.ts
@@ -2,7 +2,7 @@ import { getResource, parseJSONResponse } from '../../requester';
 import IOptions from '../../IOptions';
 import DeviceTelemetry from './DeviceTelemetry';
 import { DeviceTelemetryType } from './IDeviceTelemetry';
-import { RESOURCE as DEVICE } from "../DeviceManagement";
+import { Resources } from "../../resources";
 
 export default class DeviceTelemetryManagement {
 	public static readonly RESOURCE: string = 'telemetry';
@@ -10,7 +10,7 @@ export default class DeviceTelemetryManagement {
 	constructor(private options: IOptions) {}
 
 	public async getLatest(deviceUid: string, type: DeviceTelemetryType) {
-		const urlParts = [DEVICE, deviceUid, DeviceTelemetryManagement.RESOURCE, type, 'latest'];
+		const urlParts = [Resources.Device, deviceUid, DeviceTelemetryManagement.RESOURCE, type, 'latest'];
 		const response = await getResource(this.options, urlParts.join('/'));
 		return new DeviceTelemetry(await parseJSONResponse(response));
 	}