+++ src/RestApi/apiVersions.ts
@@ -1,0 +1,4 @@
+// TODO: Replace all occurrences of hard coded version with enum
+export enum ApiVersions {
+	V1 = 'v1',
+}